requirements.txt
# Font Detection MVP - Requirements
# Phase 1: Data Collection
# Phase 2: Model Training

# Core dependencies
pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0

# Font handling
fonttools>=4.42.0

# Image processing
opencv-python-headless>=4.8.0

# Utilities
tqdm>=4.66.0

# Phase 2: ML/AI Dependencies
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.20.0
transformers>=4.30.0

# Phase 3: Vector database
faiss-cpu>=1.7.4

# Phase 4: API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
easyocr>=1.7.0

# Visualization
matplotlib>=3.8.0
scikit-learn>=1.3.0
umap-learn>=0.5.5




